requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-allauth==0.43.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-toolbelt==0.0.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
importlib-metadata==2.0.0
oauthlib==3.1.0
Pillow==8.0.1
pipenv==2020.8.13
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==1.5
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.25.11
virtualenv==20.1.0
virtualenv-clone==0.5.4
whitenoise==5.2.0
zipp==3.4.0
